+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: serraoui <serraoui@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/14 21:17:04 by serraoui          #+#    #+#             */
-/*   Updated: 2023/12/06 12:08:31 by serraoui         ###   ########.fr       */
+/*   Updated: 2023/12/06 14:53:20 by serraoui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <stdarg.h>
 # include <unistd.h>
 
-void	ft_print_base(unsigned long n, const char *chars, int base, int *len);
+void	ft_print_base(unsigned long long n, const char *chars, int base, int *len);
 void	ft_putchar(char c, int *len);
 void	ft_putnbr(int n, int *len);
 int		ft_printf(const char *s, ...);
